Better line polygon generation

+++ yabause/src/ygl.h
@@ -71,9 +71,9 @@
 #include "vidshared.h"
 
 typedef struct {
-	int vertices[8];
-	unsigned int w;
-	unsigned int h;
+	float vertices[8];
+	int w;
+	int h;
 	int flip;
 	int priority;
 	int dst;
@@ -136,7 +136,7 @@ typedef struct {
    int prgid;
    GLuint prg;
    GLuint vertexBuffer;
-   int * quads;
+   float * quads;
    float * textcoords;
    float * vertexAttribute;
    int currentQuad;
